pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "authmcp-gateway"
version = "1.1.9"
description = "Universal Authentication Gateway for MCP (Model Context Protocol) Servers"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "AuthMCP Gateway Contributors"}
]
keywords = [
    "mcp",
    "authentication",
    "gateway",
    "oauth2",
    "jwt",
    "proxy",
    "model-context-protocol"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Security",
]

dependencies = [
    "mcp>=1.2.0",
    "starlette>=0.37.2",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "pyjwt[crypto]>=2.8.0",
    "python-dotenv>=1.0.1",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.1.0",
    "bcrypt<4.2.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/loglux/authmcp-gateway"
Documentation = "https://github.com/loglux/authmcp-gateway#readme"
Repository = "https://github.com/loglux/authmcp-gateway"
Issues = "https://github.com/loglux/authmcp-gateway/issues"

[project.scripts]
authmcp-gateway = "authmcp_gateway.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["authmcp_gateway*"]
namespaces = false

[tool.setuptools.package-data]
authmcp_gateway = [
    "py.typed",
    "templates/**/*.html",
    "static/**/*"
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=authmcp_gateway --cov-report=term-missing"
